// src/app/core/api/generated/schemas/productReviewControllerListParams.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * stallio-auto-parts API
 * OpenAPI spec version: 1.0
 */
import type { ProductReviewControllerListSortBy } from "./productReviewControllerListSortBy";

export type ProductReviewControllerListParams = {
  page?: number;
  limit?: number;
  /**
   * Search across customer name, customer email, product title, and comment
   */
  search?: string;
  review_id?: string;
  user_id?: string;
  product_id?: string;
  rating?: string;
  is_active?: string;
  customer_name?: string;
  customer_email?: string;
  product_title?: string;
  review_comment?: string;
  dateFrom?: string;
  dateTo?: string;
  sortBy?: ProductReviewControllerListSortBy;
};