// src/app/core/api/generated/schemas/productReviewControllerListSortBy.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * stallio-auto-parts API
 * OpenAPI spec version: 1.0
 */

export type ProductReviewControllerListSortBy =
  (typeof ProductReviewControllerListSortBy)[keyof typeof ProductReviewControllerListSortBy];

export const ProductReviewControllerListSortBy = {
  newest: "newest",
  oldest: "oldest",
  rating_high: "rating_high",
  rating_low: "rating_low",
  customer_asc: "customer_asc",
  customer_desc: "customer_desc",
  product_asc: "product_asc",
  product_desc: "product_desc",
} as const;